Add files via upload
This project focuses on image classification, a fundamental task in computer vision, utilizing the powerful combination of TensorFlow and Keras. Image classification involves training a machine learning model to recognize and categorize objects or scenes within images. The goal is to build a robust model capable of accurately assigning predefined labels to input images, making it applicable to various domains such as healthcare, autonomous vehicles, and security surveillance.